FORMBUILDER-BACKEND/server/backend/requirements.txt
django==4.2
djangorestframework==3.15.1
python_dotenv==1.0.1
django-cors-headers==4.3.1
psycopg2_binary==2.9.9
drf_yasg==1.21.4